8032733: Fix cast lint warnings in client libraries 8032734: Add cast lint warning to build of jdk repository

+++ jdk/src/share/classes/com/sun/media/sound/AiffFileReader.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 1999, 2013, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 1999, 2014, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -406,7 +406,7 @@ public final class AiffFileReader extends SunFileReader {
         int expon = 0;
         long hiMant = 0, loMant = 0;
         long t1, t2;
-        double HUGE = ((double)3.40282346638528860e+38);
+        double HUGE = 3.40282346638528860e+38;
 
 
         expon = dis.readUnsignedShort();
